docker_links/requirements.txt
Django==1.8.2
argparse==1.2.1
djangorestframework==3.1.3
ipdb==0.8.1
ipython==3.1.0
psycopg2==2.6
wsgiref==0.1.2
